refactor(types): consolidate type system — resolve collisions, break circular deps, absorb server types
Phase 1 — Name collision resolution:
- Rename 10 colliding types in clientTypes.ts with Client prefix
  (ClientStreamResult, ClientRetryConfig, ClientChatMessage, etc.)
- Remove 5 duplicate types from typeAliases.ts (canonical defs in common.ts)
- Rename ErrorInfo → AnalyticsErrorInfo in analytics.ts
- Update barrel index.ts and all client/* consumer files

Phase 2 — Circular dependency breaks:
- Fix cli.ts/streamTypes.ts: import EvaluationData from ./evaluation.js
  instead of ../index.js (breaks circular through barrel)
- Change configTypes.ts MCPToolRegistry to import type (breaks runtime dep)
- Move RAGConfig to types/ragTypes.ts, convert rag/types.ts to re-export

Phase 3 — Rogue type file absorption:
- Create types/serverTypes.ts with all 54 server adapter type definitions
- Convert server/types.ts to re-export shim for backward compatibility
- Add serverTypes.ts to barrel exports

Verified: 0 type errors, lint clean, full build passing, public API intact.
