+++ src/cli/commands/mcp.ts
@@ -12,14 +12,14 @@
  */
 
 import type { CommandModule, Argv } from "yargs";
-import type { UnknownRecord } from "../../lib/types/common.js";
+import type { UnknownRecord } from "../../lib/types/index.js";
 import type {
   MCPServerInfo,
   MCPStatus,
   MCPTransportType,
-} from "../../lib/types/mcpTypes.js";
+} from "../../lib/types/index.js";
 import { createExternalServerInfo } from "../../lib/utils/mcpDefaults.js";
-import type { MCPCommandArgs } from "../../lib/types/cli.js";
+import type { MCPCommandArgs } from "../../lib/types/index.js";
 import { NeuroLink } from "../../lib/neurolink.js";
 import { logger } from "../../lib/utils/logger.js";
 import chalk from "chalk";
