+++ src/cli/commands/models.ts
@@ -6,8 +6,8 @@
 
 import type { CommandModule, Argv } from "yargs";
 import type { AIProviderName } from "../../lib/constants/enums.js";
-import type { ModelsCommandArgs } from "../../lib/types/cli.js";
-import type { ModelPricing } from "../../lib/types/providers.js";
+import type { ModelsCommandArgs } from "../../lib/types/index.js";
+import type { ModelPricing } from "../../lib/types/index.js";
 import {
   ModelResolver,
   formatSearchResults,
@@ -25,7 +25,7 @@ import type {
   ModelSearchFilters,
   ModelCapabilities,
   UseCaseSuitability,
-} from "../../lib/types/modelTypes.js";
+} from "../../lib/types/index.js";
 import chalk from "chalk";
 import ora from "ora";
