+++ src/cli/loop/optionsSchema.ts
@@ -1,6 +1,6 @@
 import { AIProviderName } from "../../lib/constants/enums.js";
-import type { OptionSchema } from "../../lib/types/cli.js";
-import type { TextGenerationOptions } from "../../lib/types/generateTypes.js";
+import type { OptionSchema } from "../../lib/types/index.js";
+import type { TextGenerationOptions } from "../../lib/types/index.js";
 
 /**
  * Master schema for all text generation options.
