+++ src/cli/utils/videoFileUtils.ts
@@ -10,7 +10,7 @@
 
 import fs from "fs";
 import path from "path";
-import type { VideoGenerationResult } from "../../lib/types/multimodal.js";
+import type { VideoGenerationResult } from "../../lib/types/index.js";
 
 /**
  * Result of saving video to file
