+++ src/lib/client/index.ts
@@ -105,17 +105,17 @@ export {
 export type {
   ClientConfig,
   RequestOptions,
-  ApiResponse,
+  ClientApiResponse as ApiResponse,
   ApiError,
-  RetryConfig,
+  ClientRetryConfig as RetryConfig,
   Middleware,
   MiddlewareRequest,
   MiddlewareResponse,
-  MiddlewareContext,
+  ClientMiddlewareContext as MiddlewareContext,
   StreamCallbacks,
-  StreamEvent,
+  ClientStreamEvent as StreamEvent,
   StreamEventType,
-  StreamResult,
+  ClientStreamResult as StreamResult,
   GenerateRequestOptions,
   GenerateResponse,
   StreamRequestOptions,
@@ -126,7 +126,7 @@ export type {
   WorkflowExecuteResult,
   WorkflowInfo,
   ToolInfo,
-  ProviderStatus,
+  ClientProviderStatus as ProviderStatus,
   WebSocketOptions,
   WebSocketState,
   WebSocketMessageHandler,
@@ -163,7 +163,7 @@ export type {
   UseStreamReturn,
   UseToolsOptions,
   UseToolsReturn,
-  ChatMessage,
+  ClientChatMessage as ChatMessage,
   // AI SDK Adapter types
   NeuroLinkProviderOptions,
   ModelOptions,
@@ -297,8 +297,8 @@ export {
 
 export type {
   AuthConfig,
-  OAuth2Config,
-  TokenRefreshResult,
+  ClientOAuth2Config as OAuth2Config,
+  ClientTokenRefreshResult as TokenRefreshResult,
   // Voice types
   SpeechRecognitionResult,
   SpeechSynthesisOptions,
