+++ src/lib/client/reactHooks.tsx
@@ -42,13 +42,13 @@ import type {
   UseStreamReturn,
   UseToolsOptions,
   UseToolsReturn,
-  ChatMessage,
+  ClientChatMessage as ChatMessage,
   AgentExecuteOptions,
   AgentExecuteResult,
   WorkflowExecuteOptions,
   WorkflowExecuteResult,
   StreamCallbacks,
-  StreamEvent,
+  ClientStreamEvent as StreamEvent,
   ToolInfo,
   SpeechRecognitionInternal,
   SpeechRecognitionEventInternal,
