+++ src/lib/types/cli.ts
@@ -4,7 +4,7 @@
 
 import type { UnknownRecord, JsonValue } from "./common.js";
 import type { AnalyticsData, TokenUsage } from "./analytics.js";
-import type { EvaluationData } from "../index.js";
+import type { EvaluationData } from "./evaluation.js";
 import type { ToolCall, ToolResult } from "./tools.js";
 
 /**
